Add files via upload
finished election program. Right now the program is set to rig the election for the libertarians.
if you wish to change it, open main.py and type the party you wish to rig the election for in the 
elect.runriggedmodel method. If you wish to run the election normally, set the rigged_party parameter
in process_election_results to NONE.